pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aircraft-damage-vgg16-blip"
version = "0.1.0"
description = "Aircraft damage classification with VGG16 and BLIP-based report generation."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "gradio>=5.0",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.4",
  "tensorflow>=2.17,<2.21",
  "torch>=2.4",
  "transformers>=4.38",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
